requirements.txt
python-telegram-bot==21.4
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
websockets==12.0
tzdata==2024.1
